+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [Unreleased]
 
 - Make tool's name be the class name by default
+- Add model_tool generator to create explicit tool files for ActiveRecord models 
+- Fix issue with active_record generator not creating tool files with full implementations
 
 ## [0.2.0] - 2025-06-07